+++ NEWS.md
@@ -1,5 +1,12 @@
 # piecewiseSEM Change Log
 
+### 2016-05-18 Version 1.1.1
+  * Reduced time it takes examples to run
+  * Stop: if duplicate responses are detected in the model list
+  * Temporary bug fix: returns lowest P-value for when d-sep tests include non-linear intermediate endogenous variables, only if family is not Gaussian
+  * Bug fix: reversed order of conditioning variables so they come first 
+  * Bug fix: removed entries from the basis set that attempt to predict an interaction that does not appear in the model list
+
 ### 2016-03-10 Version 1.1.0
   * Feature addition: rudimentary plotting using `sem.plot`
   * Feature addition: support for `glmmadmb` models